add sidebar content

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('users', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->unsignedBigInteger('gold')->default(0);
+            $table->double('gold')->default(0);
             $table->unsignedBigInteger('job_id')->nullable();
             $table->unsignedBigInteger('kingdom_id')->nullable();
             $table->unsignedBigInteger('current_city_id')->nullable();
